PWGLF: stra converter modularisation
