+++ PWGLF/TableProducer/converters/stradautracksconverter.cxx
@@ -18,38 +18,10 @@ using namespace o2;
 using namespace o2::framework;
 
 // Converts V0 version 001 to 002
-struct strangeDerivedConverter {
-  Produces<aod::StraRawCents_001> straRawCents_001;
-  Produces<aod::StraRawCents_003> straRawCents_003;
+struct stradautracksconverter {
   Produces<aod::DauTrackTOFPIDs> dautracktofpids;
 
-  void processStraRawCents000to001(aod::StraRawCents_000 const& straRawCents_000)
-  {
-    for (auto& values : straRawCents_000) {
-      straRawCents_001(values.multFT0A(), values.multFT0C(), values.multFV0A(), values.multNTracksPVeta1(), 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f);
-    }
-  }
-  void processStraRawCents002to003(aod::StraRawCents_002 const& straRawCents_002)
-  {
-    for (auto& values : straRawCents_002) {
-      straRawCents_003(values.multFT0A(),
-                       values.multFT0C(),
-                       values.multFT0A(),
-                       values.multNTracksPVeta1(),
-                       0, 0,
-                       values.multNTracksITSTPC(),
-                       values.multAllTracksTPCOnly(),
-                       values.multAllTracksITSTPC(),
-                       values.multZNA(),
-                       values.multZNC(),
-                       values.multZEM1(),
-                       values.multZEM2(),
-                       values.multZPA(),
-                       values.multZPC());
-    }
-  }
-
-  void processGenerateDauTracksTOFPIDs(soa::Join<aod::V0Cores, aod::V0Extras, aod::V0TOFs> const& v0s, soa::Join<aod::CascCores, aod::CascExtras, aod::CascTOFs> const& cascs, aod::DauTrackExtras const& dauTracks)
+  void process(soa::Join<aod::V0Cores, aod::V0Extras, aod::V0TOFs> const& v0s, soa::Join<aod::CascCores, aod::CascExtras, aod::CascTOFs> const& cascs, aod::DauTrackExtras const& dauTracks)
   {
     // prepare arrays with the relevant information
     std::vector<float> lLengths, lTOFSignals, lTOFEvTimes;
@@ -84,14 +56,10 @@ struct strangeDerivedConverter {
       dautracktofpids(lTOFSignals[ii], lTOFEvTimes[ii], lLengths[ii]);
     }
   }
-
-  PROCESS_SWITCH(strangeDerivedConverter, processStraRawCents000to001, "from StraRawCents 000 to 001", false);
-  PROCESS_SWITCH(strangeDerivedConverter, processStraRawCents002to003, "from StraRawCents 002 to 003", false);
-  PROCESS_SWITCH(strangeDerivedConverter, processGenerateDauTracksTOFPIDs, "from DauTrackTOFPIDs to V0TOFs/CascTOFs", false);
 };
 
 WorkflowSpec defineDataProcessing(ConfigContext const& cfgc)
 {
   return WorkflowSpec{
-    adaptAnalysisTask<strangeDerivedConverter>(cfgc)};
+    adaptAnalysisTask<stradautracksconverter>(cfgc)};
 }
